+++ src/utils/MoviesApi.js
@@ -13,7 +13,7 @@ class MoviesApi {
     return Promise.reject(`Ошибка: ${res.status}`);
   }
 
-  getInitialMovies() {
+  getMovies() {
     return fetch(`${this._baseUrl}`, {
       method: 'GET',
       headers: this._headers,
